Refactor cmd package to pass stdio
And add first "system test" for the cmd package, using the same patterns
recently added to strangelove/lens.

This switches from plain use of fmt.Print in the cmd package, to
fmt.Fprint variants passing the command's stdin/stdout/stderr as
appropriate. The rule of thumb is that stderr is used for diagnostic
output and stdout is used for "conventional" output that one would
perhaps pipe to another process.

This also replaces the relayer.Chain type's Print method with an Sprint
method, so that callers can decide how to directly the output.