Use errgroup.WithContext where applicable (#612)
errgroup.WithContext returns a context that is canceled once the first
callback returns an error. This means, for example, if querying one
chain fails for whatever reason, the context passed to the other method
will be canceled, so the overall method will no longer block waiting for
success on one method when the whole group will fail.

We have some places where we are using an errgroup multiple times in one
scope, so I have named the derived context egCtx to hopefully clarify
the valid life of the context.

If we make a mistake using this pattern in the future, it ought to fail
in an obvious manner when a later call immediately fails with a canceled
context.

This change should merge cleanly with #611 and should not require any
new changes, as retry.Context should continue to use the outer context,
not the inner errgroup context.