Use zap logger throughout relayer package
This replaces the existing tendermint logger, and other uses of
fmt.Print*, with a provided zap logger.

I inlined several of the Log* methods on the Chain type when there were
only 2 or 3 calls.

Many calls that were checking *Chain.debug have skipped the check in
favor of a call to (*Chain).log.Debug.