Report relayer commit using debug.BuildInfo
Now that CI is building with go1.18, we can use the embedded
vcs.revision field on BuildInfo instead of including it in ldflags.
Because this is a new API in go1.18, and we are not yet ready to
officially stop builds on go1.17, include a fallback to just report
"unknown" for the commit from go1.17 builds.