Pass explicit logger to relayer package functions
Instead of relying on the logger associated with a chain, accept a
logger specifically for use in the package-level function.

By convention, ctx is always the first parameter, and then log.

Passing the logger separately also allows more flexibility in the
parameter types to these functions, which will simplify adding new tests
shortly.