Remove references to github.com/pkg/errors
All functionality in that package is provided by the standard library's
errors package.

Also run go mod tidy.

To get this to properly work with the _test directory, I had to run:

mv _test test && go mod tidy && mv test _test

So clearly the _test organization is not ideal, but it is still
convenient outside of the go mod tidy use case.