Use SendMsgsResult from all Send calls
Also log out partial successes, and treat SendMsgsResult.Error as the
canonical error to return, instead of separately returning ctx.Err at
the end of the method.

Because the error fields on SendMsgsResult are instances of multierr,
errors.Is(err, context.Canceled) should continue to work to detect a
canceled context error, regardless of whether the multierr also contains
other error values.

And remove the Succeeded field and Success method from RelayMsgs.