Prefer test logging over inline comments
I have been seeing intermittent failures in
TestRelayAllChannelsOnConnection, so I changed the inline comments to
calls to t.Log in order to make the output slightly easier to follow.

I also added one call to WaitForNBlocks where the comment suggested
there should be one; it didn't seem to make a difference in the test
flakiness.