Avoid retrying SendMessages when receiving invalid height error
There are occasions when the relayer sends messages and then is informed
by the server that the client is out of date. We cannot access a
structured error with that information, because the RPC error is
presented as a flat string.

The next best thing we can do is to detect "invalid height" in the error
message, and then stop retrying because we can be sure that no
subsequent retries will pass until the client is updated.