Merge PR #719: Add ibctest module
Running go test ./... will maintain the current behavior, but now we can
run a small version of the ibc-test-framework tests from within relayer
via the new "make ibctest" target.

As mentioned in the new README, this is not a substitute for the full
ibc-test-framework suite, which will eventually have a larger
combination of chains under test.

Note also, this does not yet add a GitHub Action to run these tests. I
am going to add that in a followup change.