Directly calculate (*CosmosProvider).TrustingPeriod
Rather than using the math package and converting back and forth to
strings, we can directly modify the raw duration and truncate it down to
the hour.

I don't know the rationale for why the trusting period is 85% of
unbonding time, or why we want it to be whole hours, but I left comments
acting like I did know the rationale.