wip: add test for relaying on many paths
This test instantiates a relayer with the event processor that relays on
two paths across three chains.

This was ported from previous WIP that attempted to add multi-path
support to the legacy processor for the relayer; multi-path support in
the legacy processor had some subtle bugs that caused the test to fail.

Now with the new event processor, the test is failing in a different
way that we need to further debug.